+++ AirTicketBookingSystem/AirTicketBookingSystem-Dao/src/main/java/cz/fi/muni/pa036/airticketbooking/entity/Administrator.java
@@ -9,7 +9,7 @@ package cz.fi.muni.pa036.airticketbooking.entity;
 public class Administrator  implements java.io.Serializable {
 
 
-     private int id;
+     private Long id;
      private Airline airline;
      private String name;
      private String password;
@@ -18,22 +18,22 @@ public class Administrator  implements java.io.Serializable {
     }
 
 	
-    public Administrator(int id, Airline airline) {
+    public Administrator(Long id, Airline airline) {
         this.id = id;
         this.airline = airline;
     }
-    public Administrator(int id, Airline airline, String name, String password) {
+    public Administrator(Long id, Airline airline, String name, String password) {
        this.id = id;
        this.airline = airline;
        this.name = name;
        this.password = password;
     }
    
-    public int getId() {
+    public Long getId() {
         return this.id;
     }
     
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
     public Airline getAirline() {
