+++ AirTicketBookingSystem/AirTicketBookingSystem-Dao/src/main/java/cz/fi/muni/pa036/airticketbooking/entity/Plane.java
@@ -12,7 +12,7 @@ import java.util.Set;
 public class Plane  implements java.io.Serializable {
 
 
-     private int id;
+     private Long id;
      private Airline airline;
      private String code;
      private String type;
@@ -25,11 +25,11 @@ public class Plane  implements java.io.Serializable {
     }
 
 	
-    public Plane(int id, Airline airline) {
+    public Plane(Long id, Airline airline) {
         this.id = id;
         this.airline = airline;
     }
-    public Plane(int id, Airline airline, String code, String type, Date creationDate, Integer maxSeats, Set flights, Set seats) {
+    public Plane(Long id, Airline airline, String code, String type, Date creationDate, Integer maxSeats, Set flights, Set seats) {
        this.id = id;
        this.airline = airline;
        this.code = code;
@@ -40,11 +40,11 @@ public class Plane  implements java.io.Serializable {
        this.seats = seats;
     }
    
-    public int getId() {
+    public Long getId() {
         return this.id;
     }
     
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
     public Airline getAirline() {
