+++ AirTicketBookingSystem/AirTicketBookingSystem-Dao/src/main/java/cz/fi/muni/pa036/airticketbooking/entity/SeatReservation.java
@@ -9,24 +9,24 @@ package cz.fi.muni.pa036.airticketbooking.entity;
 public class SeatReservation  implements java.io.Serializable {
 
 
-     private int id;
+     private Long id;
      private FlightTicket flightTicket;
      private Seat seat;
 
     public SeatReservation() {
     }
 
-    public SeatReservation(int id, FlightTicket flightTicket, Seat seat) {
+    public SeatReservation(Long id, FlightTicket flightTicket, Seat seat) {
        this.id = id;
        this.flightTicket = flightTicket;
        this.seat = seat;
     }
    
-    public int getId() {
+    public Long getId() {
         return this.id;
     }
     
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
     public FlightTicket getFlightTicket() {
